Move convert to it's own module
